+++ open-adexchange-model/src/main/java/top/openadexchange/model/Site.java
@@ -30,9 +30,9 @@ public class Site implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id(keyType = KeyType.Auto)
-    private Long id;
+    private Integer id;
 
-    private Long publisherId;
+    private Integer publisherId;
 
     private String name;
